+++ src/main/java/com/example/poems_app/xml/Seg.java
@@ -37,9 +37,6 @@ public abstract class Seg {
 		placenames = new HashSet<String>();
 	}
 	
-	@ManyToOne(cascade = {CascadeType.ALL})
-	private AuthorSection authorSection;
-	
 	@OneToOne(cascade = {CascadeType.ALL})
 	private Orig orig;
 	
@@ -115,11 +112,4 @@ public abstract class Seg {
 	@ElementCollection
 	private Set<String> placenames;
 
-	public AuthorSection getAuthorSection() {
-		return authorSection;
-	}
-
-	public void setAuthorSection(AuthorSection authorSection) {
-		this.authorSection = authorSection;
-	}	
 }
