--- src/main/java/com/example/poems_app/xml/TextGroup.java
@@ -1,35 +1,0 @@
-package com.example.poems_app.xml;
-
-import javax.persistence.CascadeType;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToOne;
-
-@Entity
-public class TextGroup {
-	
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private int id;
-	
-	@OneToOne(cascade = {CascadeType.ALL})
-    private InnerText innerText;
-
-	public InnerText getInnerText() {
-		return innerText;
-	}
-
-	public void setInnerText(InnerText innerText) {
-		this.innerText = innerText;
-	}
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-}
