+++ src/main/java/com/example/poems_app/xml/Word.java
@@ -1,0 +1,12 @@
+package com.example.poems_app.xml;
+
+import javax.persistence.Entity;
+
+@Entity
+public class Word extends Break {
+	
+	public Word() {
+		type = "WORD";
+	}
+
+}
