+++ src/test/java/com/example/poems_app/services/SolrDocumentCreatorTest.java
@@ -16,14 +16,12 @@ import org.springframework.boot.test.context.SpringBootTest;
 import com.example.poems_app.xml.AuthorSection;
 import com.example.poems_app.xml.Body;
 import com.example.poems_app.xml.ContentItem;
-import com.example.poems_app.xml.InnerText;
 import com.example.poems_app.xml.Letter;
 import com.example.poems_app.xml.Orig;
 import com.example.poems_app.xml.Part;
 import com.example.poems_app.xml.Reg;
 import com.example.poems_app.xml.Seg;
 import com.example.poems_app.xml.Text;
-import com.example.poems_app.xml.TextGroup;
 import com.example.poems_app.xml.XmlPoem;
 
 //@SpringBootTest
